+++ packages/editor/src/setup-bolt.js
@@ -6,6 +6,7 @@ import textSchema from '@bolt/components-text/text.schema.yml';
 import iconSchema from '@bolt/components-icon/icon.schema.json';
 import characterSchema from '@bolt/micro-journeys/src/character.schema';
 import connectionSchema from '@bolt/micro-journeys/src/connection.schema';
+import pathwaysSchema from '@bolt/micro-journeys/src/interactive-pathways.schema';
 import statusDialogueBarSchema from '@bolt/micro-journeys/src/status-dialogue-bar.schema';
 import svgAnimationsSchema from '@bolt/micro-journeys/src/bolt-svg-animations/svg-animations.schema';
 // @ts-ignore
@@ -459,6 +460,8 @@ export function setupBolt(editor) {
 
   registerBoltComponent({
     name: 'bolt-interactive-pathways',
+    schema: pathwaysSchema,
+    propsToTraits: ['customImageSrc', 'imageAlt'],
     category: 'Starters',
     blockTitle: 'Pathways',
     draggable: true,
@@ -622,9 +625,17 @@ export function setupBolt(editor) {
     },
   });
 
+  // The bolt-svg-animations component is only used internally by the connection component,
+  // and as a background slot for the character component. The following bit of logic removes
+  // the connection bands from the schema so that they do not appear as options when editing
+  // the svg animations behind a character.
+  const svgAnimationsSchemaForCharacter = svgAnimationsSchema;
+  svgAnimationsSchemaForCharacter.properties.animType.enum = svgAnimationsSchema.properties.animType.enum.filter(
+    item => item !== 'connectionBand' && item !== 'tripleConnectionBand',
+  );
   registerBoltComponent({
     name: 'bolt-svg-animations',
-    schema: svgAnimationsSchema,
+    schema: svgAnimationsSchemaForCharacter,
     registerBlock: false,
     propsToTraits: ['animType', 'direction', 'speed', 'theme'],
   });