+++ packages/micro-journeys/index.js
@@ -24,9 +24,6 @@ polyfillLoader.then(res => {
     /* webpackMode: 'eager', webpackChunkName: 'bolt-status-dialogue-bar' */ './src/status-dialogue-bar'
   );
   import(
-    /* webpackMode: 'eager', webpackChunkName: 'bolt-icon-group' */ './src/icon-group'
-  );
-  import(
     /* webpackMode: 'eager', webpackChunkName: 'bolt-two-character-layout' */ './src/two-character-layout'
   );
   import(